Issue #172: Move PaymentController.listAllMemberPayments to HouseMemberDocumentController and PaymentController.listAllAdminScheduledPayments to CommunityController.
This refactor includes move test to respective classes.
