+++ service/src/main/java/com/myhome/controllers/HouseMemberDocumentController.java
@@ -17,8 +17,11 @@
 package com.myhome.controllers;
 
 import com.myhome.api.MembersApi;
+import com.myhome.controllers.mapper.SchedulePaymentApiMapper;
 import com.myhome.domain.HouseMemberDocument;
+import com.myhome.model.ListMemberPaymentsResponse;
 import com.myhome.services.HouseMemberDocumentService;
+import com.myhome.services.PaymentService;
 import java.util.Optional;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.CacheControl;
@@ -40,9 +43,15 @@ import org.springframework.web.multipart.MultipartFile;
 public class HouseMemberDocumentController implements MembersApi {
 
   private final HouseMemberDocumentService houseMemberDocumentService;
+  private final PaymentService paymentService;
+  private final SchedulePaymentApiMapper schedulePaymentApiMapper;
 
-  public HouseMemberDocumentController(HouseMemberDocumentService houseMemberDocumentService) {
+  public HouseMemberDocumentController(HouseMemberDocumentService houseMemberDocumentService,
+      PaymentService paymentService,
+      SchedulePaymentApiMapper schedulePaymentApiMapper) {
     this.houseMemberDocumentService = houseMemberDocumentService;
+    this.paymentService = paymentService;
+    this.schedulePaymentApiMapper = schedulePaymentApiMapper;
   }
 
   @Override
@@ -102,5 +111,19 @@ public class HouseMemberDocumentController implements MembersApi {
     } else {
       return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
     }
+  }
+
+
+  @Override
+  public ResponseEntity<ListMemberPaymentsResponse> listAllMemberPayments(String memberId) {
+    log.trace("Received request to list all the payments for the house member with id[{}]",
+        memberId);
+
+    return paymentService.getHouseMember(memberId)
+        .map(payments -> paymentService.getPaymentsByMember(memberId))
+        .map(schedulePaymentApiMapper::memberPaymentSetToRestApiResponseMemberPaymentSet)
+        .map(memberPayments -> new ListMemberPaymentsResponse().payments(memberPayments))
+        .map(ResponseEntity::ok)
+        .orElseGet(() -> ResponseEntity.notFound().build());
   }
 }
