+++ service/src/main/java/com/myhome/controllers/mapper/SchedulePaymentApiMapper.java
@@ -19,13 +19,15 @@ package com.myhome.controllers.mapper;
 import com.myhome.controllers.dto.PaymentDto;
 import com.myhome.controllers.dto.UserDto;
 import com.myhome.controllers.request.EnrichedSchedulePaymentRequest;
-import com.myhome.controllers.request.SchedulePaymentRequest;
-import com.myhome.controllers.response.SchedulePaymentResponse;
 import com.myhome.domain.Community;
 import com.myhome.domain.HouseMember;
 import com.myhome.domain.Payment;
 import com.myhome.domain.User;
+import com.myhome.model.AdminPayment;
 import com.myhome.model.HouseMemberDto;
+import com.myhome.model.MemberPayment;
+import com.myhome.model.SchedulePaymentRequest;
+import com.myhome.model.SchedulePaymentResponse;
 import java.util.Set;
 import java.util.stream.Collectors;
 import org.mapstruct.AfterMapping;
