--- service/src/main/java/com/myhome/controllers/request/SchedulePaymentRequest.java
@@ -1,38 +1,0 @@
-/*
- * Copyright 2020 Prathab Murugan
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.myhome.controllers.request;
-
-import java.math.BigDecimal;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.Size;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Data
-public class SchedulePaymentRequest {
-  private String type;
-  @NotBlank
-  @Size(min = 5, max = 300, message = "Description should be between 5 and 300 characters")
-  private String description;
-  private boolean recurring;
-  private BigDecimal charge;
-  private String dueDate;
-  private String adminId;
-  private String memberId;
-}
